CtrlFlow: refactor RdfStorage for future adding a version working without SPARQL endpoint
